move hex/oct/bin functions to separate file
