+++ internal/c/libqb/build.mk
@@ -11,6 +11,7 @@ libqb-objs-y += $(PATH_LIBQB)/src/datetime.o
 libqb-objs-y += $(PATH_LIBQB)/src/error_handle.o
 libqb-objs-y += $(PATH_LIBQB)/src/gfs.o
 libqb-objs-y += $(PATH_LIBQB)/src/qblist.o
+libqb-objs-y += $(PATH_LIBQB)/src/hexoctbin.o
 libqb-objs-y += $(PATH_LIBQB)/src/mem.o
 libqb-objs-y += $(PATH_LIBQB)/src/math.o
 libqb-objs-y += $(PATH_LIBQB)/src/rounding.o
